requirements.txt
Flask==2.3.3
tensorflow==2.14.0
numpy
pillow
